+++ src/main/java/asp4j/solver/SolverDLV.java
@@ -2,6 +2,7 @@ package asp4j.solver;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.List;
 import org.apache.commons.io.IOUtils;
@@ -11,6 +12,10 @@ import org.apache.commons.io.IOUtils;
  * @author hbeck Apr 14, 2013
  */
 public class SolverDLV extends SolverBase {
+
+    private static final String EMPTY_LINES = "dlv output error: empty lines";
+    private static final String BAD_ANSWERSET_SYNTAX = "dlv output error: not an answer set: ";
+    private static final String MULTIPLE_LINES = "dlv output error: expected single line ouput, got: ";
 
     @Override
     protected String answerSetsSolverCommand() {
@@ -37,12 +42,12 @@ public class SolverDLV extends SolverBase {
         InputStream inputStream = exec.getInputStream();
         List<String> allLines = IOUtils.readLines(inputStream);
         if (allLines.isEmpty()) {
-            throw new IOException("dlv output error: lines empty");
+            throw new IOException(EMPTY_LINES);
         }
         List<String> answerSetLines = new ArrayList<>();
         for (String line : allLines) {
             if (!line.startsWith("{")) {
-                throw new IOException("dlv output error: not an answer set: " + line);
+                throw new IOException(BAD_ANSWERSET_SYNTAX + line);
             }
             answerSetLines.add(line);
         }
@@ -62,5 +67,22 @@ public class SolverDLV extends SolverBase {
     @Override
     protected String atomDelimiter() {
         return ", ";
+    }
+
+    @Override
+    protected boolean getBooleanQueryResult(Process exec) throws IOException, ParseException {
+        InputStream inputStream = exec.getInputStream();
+        List<String> allLines = IOUtils.readLines(inputStream);
+        if (allLines.isEmpty()) {
+            throw new IOException(EMPTY_LINES);
+        }
+        if (allLines.size() > 1) {
+            throw new IOException(MULTIPLE_LINES + allLines.size());
+        }
+        //... is bravely true.
+        //... is cautiously false.
+        String line = allLines.get(0).trim();
+        int idx = line.lastIndexOf(" ");
+        return Boolean.parseBoolean(line.substring(idx+1,line.length()-1));
     }
 }