+++ src/main/java/asp4j/solver/call/QuerySolverCall.java
@@ -3,6 +3,7 @@ package asp4j.solver.call;
 import asp4j.lang.Atom;
 import asp4j.program.Program;
 import asp4j.solver.query.Query;
+import java.util.Objects;
 
 /**
  *
@@ -24,4 +25,28 @@ public abstract class QuerySolverCall extends SolverCallBase {
         sb.append(" ").append(query.asString());
         return sb.toString();
     }
+
+    @Override
+    public int hashCode() {
+        int hash = 5;
+        hash = 37 * hash + super.hashCode();
+        hash = 37 * hash + Objects.hashCode(this.query);
+        return hash;
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (obj == null) {
+            return false;
+        }
+        if (getClass() != obj.getClass()) {
+            return false;
+        }
+        final QuerySolverCall other = (QuerySolverCall) obj;
+        if (!Objects.equals(this.query, other.query)) {
+            return false;
+        }
+        return true;
+    }    
+    
 }